Save method pointer metadata to call the function a correct way
Hook will now store metadata of the target method pointer to ensure that the method call is valid when the method pointer stores metadata (e.g. this offset when inheriting from many classes). Metadata is stored in a struct and casted with bit_cast from destination pointer.

Removed the ugly workaround of casting void* to uint64 and the to the destination signature.

Also removed the need to cast the destination pointer address to *(void**) pointer to enable it.

It is possible that this issue can also be solved by converting method pointers to __thiscall pointers, but this requires further testing.